+++ test/Integration/smoke_column_advection.jl
@@ -61,10 +61,7 @@ function boris_book_tendency!(dq, q, p, t)
         bottom = Operators.ThirdOrderOneSided(),
         top = Operators.ThirdOrderOneSided(),
     )
-    fct = Operators.FCTBorisBook(
-        bottom = Operators.FirstOrderOneSided(),
-        top = Operators.FirstOrderOneSided(),
-    )
+    fct = Operators.FCTBorisBook()
     divf2c = Operators.DivergenceF2C(
         bottom = Operators.SetValue(Geometry.WVector(FT(0))),
         top = Operators.SetValue(Geometry.WVector(FT(0))),
@@ -90,10 +87,7 @@ function zalesak_tendency!(dq, q, p, t)
         bottom = Operators.ThirdOrderOneSided(),
         top = Operators.ThirdOrderOneSided(),
     )
-    fct = Operators.FCTZalesak(
-        bottom = Operators.FirstOrderOneSided(),
-        top = Operators.FirstOrderOneSided(),
-    )
+    fct = Operators.FCTZalesak()
     divf2c = Operators.DivergenceF2C(
         bottom = Operators.SetValue(Geometry.WVector(FT(0))),
         top = Operators.SetValue(Geometry.WVector(FT(0))),
@@ -102,8 +96,7 @@ function zalesak_tendency!(dq, q, p, t)
         -divf2c(
             upwind1(w, q) + fct(
                 upwind3(w, q) - upwind1(w, q),
-                q / Δt,
-                q / Δt - divf2c(upwind1(w, q)),
+                tuple(q / Δt, q / Δt - divf2c(upwind1(w, q))),
             ),
         )
     return dq
@@ -183,11 +176,7 @@ function vanleer_tendency!(dq, q, p, t)
         bottom = Operators.SetValue(Geometry.WVector(FT(0))),
         top = Operators.SetValue(Geometry.WVector(FT(0))),
     )
-    vanleer = Operators.LinVanLeerC2F(
-        bottom = Operators.FirstOrderOneSided(),
-        top = Operators.FirstOrderOneSided(),
-        constraint = constraint,
-    )
+    vanleer = Operators.LinVanLeerC2F(constraint = constraint)
     @. dq = -divf2c(vanleer(w, q, Δt))
     return dq
 end